+++ src/modules/mo-core/data-transfer-objects/persitent-file-request.dto.ts
@@ -1,0 +1,9 @@
+import { Expose } from 'class-transformer';
+
+export class PersistentFileRequestDto {
+  @Expose()
+  root: string;
+
+  @Expose()
+  path: string;
+}