+++ src/modules/mo-core/data-transfer-objects/persitent-file.dto.ts
@@ -1,0 +1,30 @@
+import { Expose, Transform, Type } from 'class-transformer';
+
+export class PersistentFileDto {
+  @Expose()
+  name: string;
+
+  @Expose()
+  type: PersistentFileDto;
+
+  @Expose()
+  relativePath: string;
+
+  @Expose()
+  isSymbolicLink: boolean;
+
+  @Expose()
+  extension?: string;
+
+  @Expose()
+  sizeInBytes: number;
+
+  @Expose()
+  size: string;
+
+  @Expose()
+  hash: string;
+
+  @Expose()
+  mimeType?: string;
+}